pyproject.toml
[tool.poetry]
name = "academic-saas-backend"
version = "1.0.0"
description = "Academic SaaS Backend API - Multi-tenant Django REST Framework application"
authors = ["Academic SaaS Team"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
django = "4.2.9"
djangorestframework = "3.14.0"
psycopg2-binary = "2.9.9"
djangorestframework-simplejwt = "5.3.1"
python-decouple = "3.8"
dj-database-url = "2.1.0"
django-cors-headers = "4.3.1"
gunicorn = "21.2.0"
whitenoise = "6.6.0"
boto3 = "1.34.14"
django-storages = "1.14.2"
celery = "5.3.4"
redis = "5.0.1"
django-celery-beat = "2.5.0"
django-filter = "23.5"
drf-spectacular = "0.27.0"
pillow = "10.2.0"
python-dateutil = "2.8.2"
psutil = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-django = "^4.0"
black = "^23.0"
flake8 = "^6.0"
isort = "^5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88